+++ scripts/macros/2024/classFeatures/artificer/battleReady.js
@@ -1,0 +1,6 @@
+import { battleReady as battleReadyLegacy } from '../../../../legacyMacros.js';
+export let battleReady = {
+    ...battleReadyLegacy,
+    version: '1.5.17',
+    rules: 'modern'
+};